print ints instead of floats -- to reduce code size for rbbb

+++ pilot/modules/attitude/attitude_test.c
@@ -23,11 +23,7 @@ int main(){
         vector_t a = accel_get();
         vector_t pv = attitude(g, a);
         
-//        sprintf(temp, "gx=%+1.2f, gy=%+1.2f, gz=%+1.2f\n\r", g.x, g.y, g.z);
-//        serial_write_s(temp);
-//        sprintf(temp, "ax=%+1.2f, ay=%+1.2f, az=%+1.2f\n\r", a.x, a.y, a.z);
-//        serial_write_s(temp);
-        sprintf(temp, " x=%+1.2f,  y=%+1.2f,  z=%+1.2f\n\r", pv.x * 57.2957795, pv.y * 57.2957795, pv.z * 57.2957795);
+        sprintf(temp, " x=%+d,  y=%+d,  z=%+d\n\r", (int) (pv.x * 57.2957795), (int) (pv.y * 57.2957795), (int) (pv.z * 57.2957795));
         serial_write_s(temp);
     }
 }
